chore: split release notes into single files
Because we dont "merge back" the releases to the master branch anymore, the release info has to be manually "synced". This is much easier to do with 1 file per relase, as it leads to less merge conflicts (and makes them more readable if they would still occur).

To keep a better overview, the latest release is now also mentioned in the README.

format:

- files like the git project does https://github.com/git/git/tree/master/Documentation/RelNotes
- content in Markdown like before, but now in standard format so readable on github
- rest of data stays structured/YAML (also readable/formatted on github)

quirks:

- for deployment there is still a single YAML file generated, no changes in usage to the previous, manually managed file in repo
- first 2 RCs were not correctly named for sorting according to semver. change `rc` to `0.rc` so the list has correct order when auto-generated